move root scanning code to its own header
so as to use it from MMTk or what ever
