+++ src/gctools/memoryManagement.cc
@@ -47,6 +47,7 @@ THE SOFTWARE.
 #include <clasp/gctools/gcFunctions.h>
 #include <clasp/gctools/snapshotSaveLoad.h>
 #include <clasp/gctools/memoryManagement.h>
+#include <clasp/gctools/roots.h>
 #include <clasp/core/mpPackage.h>
 #include <clasp/llvmo/llvmoExpose.h>
 #include <clasp/llvmo/code.h>
@@ -456,46 +457,6 @@ void FinishAssingingBuiltinStamps() {
 
 namespace gctools {
 
-/* Walk all of the global (not per-thread) roots,
- * passing the address of each root */
-template <std::invocable<Tagged*> RootWalkCallback>
-void walkGlobalRoots(RootWalkCallback&& callback) {
-  // luckily, just the one god object references everything else
-  callback((Tagged*)&_lisp);
-};
-
-template <std::invocable<Tagged*> ThreadWalkCallback>
-void walkThreadRoots(ThreadWalkCallback&& callback) {
-  // NOTE we don't need the threads mutex since we must have stopped the world.
-  for (auto cur : _lisp->_Roots._ActiveThreads) {
-    mp::Process_sp proc = core::oCar(cur).as_assert<mp::Process_O>();
-    core::ThreadLocalState* tls = proc->_ThreadInfo;
-    tls->walkRoots(callback);
-    tls->walkVMStack(callback);
-    tls->walkControlStack([&](Tagged* tp) {
-      // The control stack we have to walk conservatively.
-      switch(ptag(*tp)) {
-      case general_tag: {
-        Header_s* header = (Header_s*)GeneralPtrToHeaderPtr(untag_object((void*)*tp));
-        if (header->isValidGeneralObject())
-          callback(tp);
-      } break;
-      case cons_tag: {
-        ConsHeader_s* header = (ConsHeader_s*)ConsPtrToHeaderPtr(untag_object((void*)*tp));
-        if (header->isValidConsObject())
-          callback(tp);
-      } break;
-      default: callback(tp);
-      }
-    });
-  }
-}
-
-template <std::invocable<Tagged*> WalkCallback>
-void walkRoots(WalkCallback&& callback) {
-  walkGlobalRoots(callback);
-  walkThreadRoots(callback);
-}
 
 static void mw_obj_scan(core::General_O* client,
                         std::stack<std::pair<Tagged, Tagged>>& markStack) {
